backend/requirements.txt
fastapi
uvicorn
motor
pymongo
python-dotenv
bcrypt
PyJWT
firebase-admin
python-multipart
slowapi
networkx
google-genai
pytest
httpx